+++ dropwizard-guicey/src/test/groovy/ru/vyarus/dropwizard/guice/config/disable/DisableDropwizardBundleTest.groovy
@@ -1,4 +1,4 @@
-package ru.vyarus.dropwizard.guice.config
+package ru.vyarus.dropwizard.guice.config.disable
 
 import com.google.inject.Inject
 import io.dropwizard.core.Application